Multiple updates
GAM has been updated to address the issues addressed in the article: "Discontinuing support for JSON-RPC and Global HTTP Batch Endpoints" at https://developers.googleblog.com/.

GAM has been updated to use the `patch` method with `gam update calendar` so that the values of parameters `selected` and `hidden` are preserved.

Added the ability to transfer all of a user's assets: drive files, calendars and Google+ information, in a single command.

Added a command to display transfer requests in a formatted form.